requirements.txt
blue
ruff
fastapi[standard]
sqlalchemy
pytest 
pytest-cov
httpx
email-validator
pytest-asyncio
passlib
pydantic-settings
bcrypt==4.0.1
aiosqlite
pre-commit
mypy
bandit
PyJWT
types-PyJWT
asyncpg